+++ api/lib/application/pole-emploi/pole-emploi-controller.js
@@ -1,7 +1,7 @@
 const usecases = require('../../domain/usecases');
 const tokenService = require('../../domain/services/token-service');
 const userRepository = require('../../infrastructure/repositories/user-repository');
-const authenticationService = require('../../../lib/domain/services/authentication-service');
+const poleEmploiAuthenticationService = require('../../../lib/domain/services/pole-emploi-authentication-service');
 
 module.exports = {
   async createUser(request, h) {
@@ -27,7 +27,7 @@ module.exports = {
   },
 
   async getAuthUrl(request, h) {
-    const result = authenticationService.getPoleEmploiAuthUrl({
+    const result = poleEmploiAuthenticationService.getAuthUrl({
       redirectUri: request.query['redirect_uri'],
     });
     return h.response(result).code(200);
