+++ api/lib/domain/services/authentication-service.js
@@ -1,17 +1,4 @@
-const querystring = require('querystring');
-const get = require('lodash/get');
-
-const settings = require('../../config');
-const httpAgent = require('../../infrastructure/http/http-agent');
-
-const { GeneratePoleEmploiTokensError, GenerateCnavTokensError } = require('../errors');
-
-const PoleEmploiTokens = require('../models/PoleEmploiTokens');
-const CnavTokens = require('../models/CnavTokens');
-
 const encryptionService = require('./encryption-service');
-const tokenService = require('./token-service');
-const { v4: uuidv4 } = require('uuid');
 
 async function getUserByUsernameAndPassword({ username, password, userRepository }) {
   const foundUser = await userRepository.getByUsernameOrEmailWithRolesAndPassword(username);
@@ -25,150 +12,6 @@ async function getUserByUsernameAndPassword({ username, password, userRepository
   return foundUser;
 }
 
-async function exchangePoleEmploiCodeForTokens({ code, redirectUri }) {
-  const data = {
-    client_secret: settings.poleEmploi.clientSecret,
-    grant_type: 'authorization_code',
-    code,
-    client_id: settings.poleEmploi.clientId,
-    redirect_uri: redirectUri,
-  };
-
-  const tokensResponse = await httpAgent.post({
-    url: settings.poleEmploi.tokenUrl,
-    payload: querystring.stringify(data),
-    headers: { 'content-type': 'application/x-www-form-urlencoded' },
-  });
-
-  if (!tokensResponse.isSuccessful) {
-    const errorMessage = _getErrorMessage(tokensResponse.data);
-    throw new GeneratePoleEmploiTokensError(errorMessage, tokensResponse.code);
-  }
-
-  return new PoleEmploiTokens({
-    accessToken: tokensResponse.data['access_token'],
-    idToken: tokensResponse.data['id_token'],
-    expiresIn: tokensResponse.data['expires_in'],
-    refreshToken: tokensResponse.data['refresh_token'],
-  });
-}
-
-async function exchangeCnavCodeForTokens({ code, redirectUri }) {
-  const data = {
-    client_secret: settings.cnav.clientSecret,
-    grant_type: 'authorization_code',
-    code,
-    client_id: settings.cnav.clientId,
-    redirect_uri: redirectUri,
-  };
-
-  const tokensResponse = await httpAgent.post({
-    url: settings.cnav.tokenUrl,
-    payload: querystring.stringify(data),
-    headers: { 'content-type': 'application/x-www-form-urlencoded' },
-  });
-
-  if (!tokensResponse.isSuccessful) {
-    const errorMessage = _getErrorMessage(tokensResponse.data);
-    throw new GenerateCnavTokensError(errorMessage, tokensResponse.code);
-  }
-
-  return new CnavTokens({
-    accessToken: tokensResponse.data['access_token'],
-    idToken: tokensResponse.data['id_token'],
-    expiresIn: tokensResponse.data['expires_in'],
-    refreshToken: tokensResponse.data['refresh_token'],
-  });
-}
-
-async function getPoleEmploiUserInfo(idToken) {
-  const { given_name, family_name, nonce, idIdentiteExterne } = await tokenService.extractPayloadFromPoleEmploiIdToken(
-    idToken
-  );
-
-  return {
-    firstName: given_name,
-    lastName: family_name,
-    externalIdentityId: idIdentiteExterne,
-    nonce,
-  };
-}
-
-async function getCnavUserInfo(idToken) {
-  const { given_name, family_name, nonce, sub } = await tokenService.extractClaimsFromCnavIdToken(idToken);
-
-  return {
-    firstName: given_name,
-    lastName: family_name,
-    externalIdentityId: sub,
-    nonce,
-  };
-}
-
-function getPoleEmploiAuthUrl({ redirectUri }) {
-  const redirectTarget = new URL(`${settings.poleEmploi.authUrl}`);
-  const state = uuidv4();
-  const nonce = uuidv4();
-  const clientId = settings.poleEmploi.clientId;
-  const params = [
-    { key: 'state', value: state },
-    { key: 'nonce', value: nonce },
-    { key: 'realm', value: '/individu' },
-    { key: 'client_id', value: clientId },
-    { key: 'redirect_uri', value: redirectUri },
-    { key: 'response_type', value: 'code' },
-    {
-      key: 'scope',
-      value: `application_${clientId} api_peconnect-individuv1 openid profile serviceDigitauxExposition api_peconnect-servicesdigitauxv1`,
-    },
-  ];
-
-  params.forEach(({ key, value }) => redirectTarget.searchParams.append(key, value));
-
-  return { redirectTarget: redirectTarget.toString(), state, nonce };
-}
-
-function getCnavAuthUrl({ redirectUri }) {
-  const redirectTarget = new URL(settings.cnav.authUrl);
-  const state = uuidv4();
-  const nonce = uuidv4();
-  const clientId = settings.cnav.clientId;
-  const params = [
-    { key: 'state', value: state },
-    { key: 'nonce', value: nonce },
-    { key: 'client_id', value: clientId },
-    { key: 'redirect_uri', value: redirectUri },
-    { key: 'response_type', value: 'code' },
-    {
-      key: 'scope',
-      value: 'openid profile',
-    },
-  ];
-
-  params.forEach(({ key, value }) => redirectTarget.searchParams.append(key, value));
-
-  return { redirectTarget: redirectTarget.toString(), state, nonce };
-}
-
-function _getErrorMessage(data) {
-  let message;
-
-  if (data instanceof String) {
-    message = data;
-  } else {
-    const error = get(data, 'error', '');
-    const error_description = get(data, 'error_description', '');
-    message = `${error} ${error_description}`;
-  }
-  return message.trim();
-}
-
 module.exports = {
-  exchangePoleEmploiCodeForTokens,
-  getPoleEmploiUserInfo,
-  getPoleEmploiAuthUrl,
   getUserByUsernameAndPassword,
-  getCnavAuthUrl,
-  getCnavUserInfo,
-  exchangeCnavCodeForTokens,
 };
