+++ api/lib/domain/usecases/create-user-from-pole-emploi.js
@@ -10,13 +10,13 @@ module.exports = async function createUserFromPoleEmploi({
   authenticationMethodRepository,
   poleEmploiTokensRepository,
   userToCreateRepository,
-  authenticationService,
+  poleEmploiAuthenticationService,
 }) {
   const poleEmploiTokens = await poleEmploiTokensRepository.getByKey(authenticationKey);
   if (!poleEmploiTokens) {
     throw new AuthenticationKeyForPoleEmploiTokenExpired();
   }
-  const userInfo = await authenticationService.getPoleEmploiUserInfo(poleEmploiTokens.idToken);
+  const userInfo = await poleEmploiAuthenticationService.getUserInfo(poleEmploiTokens.idToken);
 
   if (!userInfo.firstName || !userInfo.lastName || !userInfo.externalIdentityId) {
     logger.error(`Un des champs obligatoires n'a pas été renvoyé par /userinfo: ${JSON.stringify(userInfo)}.`);
