+++ api/lib/domain/usecases/index.js
@@ -57,6 +57,7 @@ const dependencies = {
   challengeRepository: require('../../infrastructure/repositories/challenge-repository'),
   challengeForPixAutoAnswerRepository: require('../../infrastructure/repositories/challenge-for-pix-auto-answer-repository'),
   cleaCertificationResultRepository: require('../../infrastructure/repositories/clea-certification-result-repository'),
+  cnavAuthenticationService: require('../../domain/services/cnav-authentication-service'),
   competenceEvaluationRepository: require('../../infrastructure/repositories/competence-evaluation-repository'),
   competenceMarkRepository: require('../../infrastructure/repositories/competence-mark-repository'),
   competenceRepository: require('../../infrastructure/repositories/competence-repository'),
@@ -102,6 +103,7 @@ const dependencies = {
   pixPlusDroitMaitreCertificationResultRepository: require('../../infrastructure/repositories/pix-plus-droit-maitre-certification-result-repository'),
   pixPlusDroitExpertCertificationResultRepository: require('../../infrastructure/repositories/pix-plus-droit-expert-certification-result-repository'),
   placementProfileService: require('../../domain/services/placement-profile-service'),
+  poleEmploiAuthenticationService: require('../../domain/services/pole-emploi-authentication-service'),
   poleEmploiSendingRepository: require('../../infrastructure/repositories/pole-emploi-sending-repository'),
   poleEmploiTokensRepository: require('../../infrastructure/repositories/pole-emploi-tokens-repository'),
   cnavTokensRepository: require('../../infrastructure/repositories/cnav-tokens-repository'),
