+++ api/tests/unit/domain/usecases/create-user-from-pole-emploi_test.js
@@ -24,7 +24,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
   let authenticationMethodRepository;
   let poleEmploiTokensRepository;
   let userToCreateRepository;
-  let authenticationService;
+  let poleEmploiAuthenticationService;
 
   const now = new Date('2021-01-02');
 
@@ -49,8 +49,8 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
       create: sinon.stub(),
     };
 
-    authenticationService = {
-      getPoleEmploiUserInfo: sinon.stub(),
+    poleEmploiAuthenticationService = {
+      getUserInfo: sinon.stub(),
     };
   });
 
@@ -69,7 +69,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
       authenticationMethodRepository,
       poleEmploiTokensRepository,
       userToCreateRepository,
-      authenticationService,
+      poleEmploiAuthenticationService,
     });
 
     // then
@@ -96,7 +96,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         externalIdentityId: 'externalIdentityId',
         nonce: 'nonce',
       };
-      authenticationService.getPoleEmploiUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
+      poleEmploiAuthenticationService.getUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
 
       authenticationMethodRepository.findOneByExternalIdentifierAndIdentityProvider
         .withArgs({
@@ -124,7 +124,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         authenticationMethodRepository,
         poleEmploiTokensRepository,
         userToCreateRepository,
-        authenticationService,
+        poleEmploiAuthenticationService,
       });
 
       // then
@@ -164,7 +164,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         nonce: 'nonce',
       };
 
-      authenticationService.getPoleEmploiUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
+      poleEmploiAuthenticationService.getUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
 
       sinon.stub(logger, 'error');
 
@@ -174,7 +174,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         authenticationMethodRepository,
         poleEmploiTokensRepository,
         userToCreateRepository,
-        authenticationService,
+        poleEmploiAuthenticationService,
       });
 
       // then
@@ -206,7 +206,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         externalIdentityId: 'externalIdentityId',
         nonce: 'nonce',
       };
-      authenticationService.getPoleEmploiUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
+      poleEmploiAuthenticationService.getUserInfo.withArgs(poleEmploiTokens.idToken).resolves(decodedUserInfo);
 
       const authenticationMethod = domainBuilder.buildAuthenticationMethod.withPoleEmploiAsIdentityProvider({ userId });
       authenticationMethodRepository.findOneByExternalIdentifierAndIdentityProvider
@@ -222,7 +222,7 @@ describe('Unit | UseCase | create-user-from-pole-emploi', function () {
         authenticationMethodRepository,
         poleEmploiTokensRepository,
         userToCreateRepository,
-        authenticationService,
+        poleEmploiAuthenticationService,
       });
 
       // then
